Minor fixes
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@3094 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/had_amr3d/amr_c/logging.cpp
@@ -70,13 +70,13 @@ namespace hpx { namespace components { namespace amr { namespace server
         if ( logcode == 0 ) {
           if (fmod(val.timestep_,par->output) < 1.e-6 && val.level_ >= par->output_level) {
             for (i=0;i<val.granularity;i++) {
-              x.push_back(val.x_[0]+i*par->dx0/pow(2.0,val.level_));
+              x.push_back(val.x_[0]+i*par->dx0/pow(2.0,(int)val.level_));
             }
             for (i=0;i<val.granularity;i++) {
-              x.push_back(val.y_[0]+i*par->dx0/pow(2.0,val.level_));
+              x.push_back(val.y_[0]+i*par->dx0/pow(2.0,(int)val.level_));
             }
             for (i=0;i<val.granularity;i++) {
-              x.push_back(val.z_[0]+i*par->dx0/pow(2.0,val.level_));
+              x.push_back(val.z_[0]+i*par->dx0/pow(2.0,(int)val.level_));
             }
             for (i=0;i<val.granularity*val.granularity*val.granularity;i++) {
               phi.push_back(val.value_[i].phi[0][0]);
